[WFLY-20231] Rename the messaging-activemq injection package name to ensure the package name is unique.
https://issues.redhat.com/browse/WFLY-20231
