+++ messaging-activemq/injection/src/test/java/org/wildfly/extension/messaging/activemq/injection/deployment/AbstractJMSContextTestCase.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.wildfly.extension.messaging.activemq.deployment.injection;
+package org.wildfly.extension.messaging.activemq.injection.deployment;
 
 import java.util.LinkedList;
 import java.util.List;
